refactor(orchestrator): extract ToolLoopRunner + ToolLoopResumer
R2 step 5, completing R2. AgentOrchestrator: 2,725 -> 1,163 lines.

ToolLoopRunner owns the live loop: drive the model, gate and execute the
tools it asks for, meter cost and context budget, pause when a gated call
needs a human. ToolLoopResumer owns the other way in: replay the transcript,
apply the human's verdicts call by call, hand back to the same loop.

The two share one execution pipeline by construction. Every approved call
goes through ToolLoopRunner.executeSingleToolCallResult and the continuation
through runToolCallLoop, so rate limits, cache hits, cost charges, tenant
budgets and LAZY activation behave identically whether a call was requested
by the model or approved by a person. Already true before this; now stated in
both classes, because a divergence there is the kind nobody notices until an
audit asks why approved calls were priced differently.

Moved by exact line ranges rather than retyped: ~470 lines of live
LLM-execution code, where a transcription slip in runToolCallLoop would be a
silent behavioural change on the path every agent turn takes. Only the call
sites that had to change were rewritten -- seven in the runner, plus the
resumer's calls back into the facade and the runner.

ToolLoopResumer holds a back-reference to AgentOrchestrator for the two
things resume needs from the facade: buildToolSetup, so a resumed turn
rebuilds tooling through the same provider assembly the live path used, and
collectEnabledTools for the history-rebuild fallback. Constructor only stores
it, same as MemberTurnExecutor in R1.

Bare-token sweep first: activateDiscoveredTools (28 refs), toJson (10) and
restoreActiveSpecs (9) are reached via getDeclaredMethod, auditOutcomeUnknown
is a direct instance call. All keep declared delegators.

1,004 tests green. Checkstyle down to 7 violations; the only FileLength one
left is ConversationService, which is R3.
