+++ src/main/java/ai/labs/eddi/modules/llm/impl/AgentOrchestrator.java
@@ -12,9 +12,6 @@ import ai.labs.eddi.configs.hitl.model.ToolApprovalsConfig;
 import ai.labs.eddi.engine.hitl.tools.ChatTranscriptCodec;
 import ai.labs.eddi.engine.hitl.tools.IHitlToolJournalStore;
 import ai.labs.eddi.engine.hitl.tools.ToolApprovalGate;
-import ai.labs.eddi.engine.hitl.tools.ToolApprovalRequiredException;
-import ai.labs.eddi.engine.hitl.tools.ToolApprovalRules;
-import ai.labs.eddi.engine.lifecycle.model.ToolCallDecision;
 import ai.labs.eddi.engine.memory.model.PendingToolCallBatch;
 import ai.labs.eddi.configs.workflows.IRestWorkflowStore;
 import ai.labs.eddi.datastore.serialization.IJsonSerialization;
@@ -38,11 +35,7 @@ import ai.labs.eddi.modules.llm.impl.orchestration.ToolContextBudget;
 import ai.labs.eddi.modules.llm.tools.spi.ToolAssemblyContext;
 import ai.labs.eddi.modules.llm.tools.spi.ToolContribution;
 import ai.labs.eddi.modules.llm.tools.spi.ToolSourceRegistry;
-import ai.labs.eddi.modules.llm.tools.ToolCacheService;
-import ai.labs.eddi.modules.llm.tools.ToolCostTracker;
 import ai.labs.eddi.modules.llm.tools.ToolExecutionService;
-import ai.labs.eddi.modules.llm.tools.ToolInvocation;
-import ai.labs.eddi.modules.llm.tools.ToolNameResolver;
 import ai.labs.eddi.modules.llm.tools.impl.*;
 import dev.langchain4j.agent.tool.ToolExecutionRequest;
 import dev.langchain4j.agent.tool.ToolSpecification;
@@ -50,8 +43,6 @@ import dev.langchain4j.agent.tool.ToolSpecifications;
 import dev.langchain4j.data.message.*;
 import dev.langchain4j.model.TokenCountEstimator;
 import dev.langchain4j.model.chat.ChatModel;
-import dev.langchain4j.model.chat.request.ChatRequest;
-import dev.langchain4j.model.chat.response.ChatResponse;
 import dev.langchain4j.model.output.TokenUsage;
 import dev.langchain4j.service.tool.ToolExecutor;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -62,13 +53,9 @@ import org.jboss.logging.Logger;
 import static ai.labs.eddi.utils.LogSanitizer.sanitize;
 
 import ai.labs.eddi.engine.tenancy.TenantQuotaService;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
-
-import static ai.labs.eddi.utils.RuntimeUtilities.isNullOrEmpty;
 
 /**
  * Executes the tool-calling agent loop against a ChatModel.
@@ -109,7 +96,7 @@ class AgentOrchestrator implements IAgentOrchestrator {
     /** Default max pauses per turn when the config does not specify one. */
     private static final int DEFAULT_MAX_PAUSES_PER_TURN = 3;
     /** Default transcript serialization cap (bytes). */
-    private static final int DEFAULT_TRANSCRIPT_MAX_BYTES = PendingToolCallBatch.TRANSCRIPT_MAX_BYTES_DEFAULT;
+    static final int DEFAULT_TRANSCRIPT_MAX_BYTES = PendingToolCallBatch.TRANSCRIPT_MAX_BYTES_DEFAULT;
     /** Approver-facing pause reason cap (chars). */
     private static final int PAUSE_REASON_MAX_CHARS = 500;
 
@@ -153,7 +140,7 @@ class AgentOrchestrator implements IAgentOrchestrator {
      * {@code @ApplicationScoped} bean in PR #604 — that makes
      * {@code @ConfigProperty} viable for *instance* state, but not for this one.)
      */
-    private static final boolean BUDGET_ENFORCE_DEFAULT = resolveBudgetEnforceDefault();
+    static final boolean BUDGET_ENFORCE_DEFAULT = resolveBudgetEnforceDefault();
 
     private static boolean resolveBudgetEnforceDefault() {
         try {
@@ -181,7 +168,7 @@ class AgentOrchestrator implements IAgentOrchestrator {
      * @param task
      *            the configured LLM task about to run
      */
-    private static void warnAboutUnenforcedBudgets(LlmConfiguration.Task task) {
+    static void warnAboutUnenforcedBudgets(LlmConfiguration.Task task) {
         if (task.getMaxBudgetPerConversation() == null || task.getEnforceBudget() != null || BUDGET_ENFORCE_DEFAULT) {
             return;
         }
@@ -210,6 +197,14 @@ class AgentOrchestrator implements IAgentOrchestrator {
      * 4).
      */
     private final ToolApprovalGateSupport gateSupport = new ToolApprovalGateSupport(chatTranscriptCodec);
+
+    /**
+     * The live tool-calling loop, extracted to {@link ToolLoopRunner} (R2 step 5).
+     */
+    private final ToolLoopRunner toolLoopRunner;
+
+    /** The HITL resume path, extracted to {@link ToolLoopResumer} (R2 step 5). */
+    private final ToolLoopResumer toolLoopResumer;
 
     // Built-in tools
     private final CalculatorTool calculatorTool;
@@ -271,7 +266,7 @@ class AgentOrchestrator implements IAgentOrchestrator {
 
     // HITL tool-approval resume dependencies (Task 9)
     private final IHitlToolJournalStore journalStore;
-    private final ConversationHistoryBuilder conversationHistoryBuilder;
+    final ConversationHistoryBuilder conversationHistoryBuilder;
 
     /**
      * The SAME estimator factory {@code LlmTask} uses to window conversation
@@ -350,6 +345,9 @@ class AgentOrchestrator implements IAgentOrchestrator {
         this.httpCallToolsProvider = new HttpCallToolsProvider(restAgentStore, restWorkflowStore, resourceClientLibrary,
                 apiCallExecutor, jsonSerialization, memoryItemConverter);
         this.mcpToolsProvider = new McpToolsProvider(restAgentStore, restWorkflowStore, resourceClientLibrary, mcpToolProviderManager);
+        this.toolLoopRunner = new ToolLoopRunner(toolExecutionService, toolResponseTruncator, tenantQuotaService,
+                memorySnapshotService, toolApprovalGate, gateSupport, toolContextBudgetGuard);
+        this.toolLoopResumer = new ToolLoopResumer(this, toolLoopRunner, gateSupport, chatTranscriptCodec, journalStore);
         this.builtinToolsProvider = new BuiltinToolsProvider(calculatorTool, dateTimeTool, webSearchTool,
                 dataFormatterTool, webScraperTool, textSummarizerTool, pdfReaderTool, weatherTool,
                 fetchToolResponsePageTool);
@@ -498,238 +496,28 @@ class AgentOrchestrator implements IAgentOrchestrator {
         return resumeToolLoop(chatModel, task, memory, batch, decision, toolHitlEnabled, JsonResponseFormatPolicy.DISABLED);
     }
 
-    /**
-     * As
-     * {@link #resumeToolLoop(ChatModel, LlmConfiguration.Task, IConversationMemory, PendingToolCallBatch, HitlDecision, boolean)},
-     * but carrying the JSON response-format policy so the continuation's requests
-     * match what the live loop would have sent.
-     */
+    /** @see ToolLoopResumer#resumeToolLoop */
     @Override
     public ExecutionResult resumeToolLoop(ChatModel chatModel, LlmConfiguration.Task task, IConversationMemory memory,
                                           PendingToolCallBatch batch,
                                           HitlDecision decision, boolean toolHitlEnabled,
                                           JsonResponseFormatPolicy jsonPolicy)
             throws LifecycleException {
-
-        String conversationId = memory.getConversationId();
-        String pauseEpoch = batch.getPauseEpoch();
-        List<Map<String, Object>> trace = new ArrayList<>();
-
-        // Tool-cost baseline, snapshotted BEFORE the verdict loop below — that loop
-        // runs every human-approved gated call through executeToolWrapped, which
-        // charges the conversation's cost tracker. A baseline taken after it would
-        // already contain those charges, so the delta reported as toolCostUsd (and
-        // hence the audit ledger's dollar figure) would exclude exactly the calls a
-        // human explicitly approved. The live path (executeWithTools) snapshots
-        // before any tool runs; this is the same contract.
-        double toolCostBefore = conversationToolCost(conversationId);
-
-        // ── Step 2: rebuild tooling via the shared setup (SAME as the live path) ──
-        ToolSetup setup = buildToolSetup(task, memory);
-        boolean isLazy = task.getToolLoadingStrategy() == LlmConfiguration.ToolLoadingStrategy.LAZY;
-
-        // Restore the active-spec surface. For LAZY, reactivate exactly the specs that
-        // were active at pause time (activatedToolNames); otherwise the full set.
-        List<ToolSpecification> activeSpecs = restoreActiveSpecs(setup, isLazy, batch.getActivatedToolNames());
-
-        // ── Step 1: reconstitute the transcript (primary) or fall back to rebuild ──
-        List<ChatMessage> currentMessages = new ArrayList<>();
-        boolean transcriptRestored = false;
-        if (!batch.isTranscriptOmitted() && batch.getChatTranscriptJson() != null) {
-            try {
-                currentMessages.addAll(chatTranscriptCodec.deserialize(batch.getChatTranscriptJson()));
-                transcriptRestored = true;
-            } catch (ChatTranscriptCodec.TranscriptCodecException e) {
-                LOGGER.warnf("HITL resume: transcript restore failed (%s) — falling back to history rebuild for conversation '%s'",
-                        e.getMessage(), sanitize(conversationId));
-            }
-        }
-        if (!transcriptRestored) {
-            currentMessages.addAll(fallbackRebuildMessages(task, memory, batch));
-        }
-        trace.add(Map.of("type", "hitl_resume", "transcriptRestored", transcriptRestored));
-
-        // ── Step 3: apply verdicts in batch order ──
-        Set<String> clearedCallIds = new HashSet<>();
-        Map<String, ToolExecutor> toolExecutors = setup.toolExecutors();
-        HitlDecision.HitlVerdict topVerdict = decision.getVerdict();
-        Map<String, ToolCallDecision> perCall = decision.getToolDecisions() != null ? decision.getToolDecisions() : Map.of();
-
-        // Per-request execution controls (mirrors the live loop).
-        boolean enableRateLimiting = task.getEnableRateLimiting() != null ? task.getEnableRateLimiting() : true;
-        boolean enableCaching = task.getEnableToolCaching() != null ? task.getEnableToolCaching() : true;
-        boolean enableCostTracking = task.getEnableCostTracking() != null ? task.getEnableCostTracking() : true;
-        int defaultRateLimit = task.getDefaultRateLimit() != null ? task.getDefaultRateLimit() : 100;
-        Map<String, Integer> toolRateLimits = task.getToolRateLimits();
-        Map<String, String> toolCanonicalNames = setup.toolCanonicalNames();
-        Double maxBudget = task.getMaxBudgetPerConversation();
-        List<ToolSpecification> builtInSpecs = setup.builtInSpecs();
-
-        for (PendingToolCallBatch.PendingToolCall c : batch.getCalls()) {
-            ToolCallDecision cd = perCall.get(c.getCallId());
-            HitlDecision.HitlVerdict verdict = cd != null && cd.getVerdict() != null ? cd.getVerdict() : topVerdict;
-            String note = cd != null ? cd.getNote() : decision.getNote();
-            String amended = cd != null ? cd.getAmendedArguments() : null;
-
-            if (verdict == HitlDecision.HitlVerdict.REJECTED) {
-                currentMessages.add(ToolExecutionResultMessage.from(rebuiltRequest(c), rejectionEnvelope(c.getToolName(), note)));
-                trace.add(Map.of("type", "hitl_rejected", "tool", c.getToolName(), "callId", c.getCallId()));
-                continue;
-            }
-
-            // APPROVED (top-level default or per-call) below.
-
-            // Truncated raw args can't be honestly executed (the approver never saw the
-            // full args and the raw bytes are incomplete). Validation already blocks
-            // amendments on these — approving yields an honest non-execution.
-            if (c.isArgsTruncated()) {
-                currentMessages.add(ToolExecutionResultMessage.from(rebuiltRequest(c),
-                        "{\"status\":\"NOT_EXECUTED\",\"reason\":\"arguments exceeded the persistable size cap\"}"));
-                trace.add(Map.of("type", "hitl_not_executed", "tool", c.getToolName(), "callId", c.getCallId()));
-                continue;
-            }
-
-            // Journal protocol — at-most-once across crashes/re-approvals.
-            if (journalStore.tryClaim(conversationId, pauseEpoch, c.getCallId(), c.getToolName(), decision.getDecidedBy())) {
-                String args = amended != null ? amended : c.getArgumentsRaw();
-                ToolExecutionRequest req = rebuiltRequest(c, args);
-                // Full per-request pipeline (checkpoint, budget, executeToolWrapped,
-                // truncation, trace). Its own auto-checkpoint fires ONLY here.
-                String result = executeSingleToolCallResult(req, memory, trace, toolExecutors, toolRateLimits,
-                        toolCanonicalNames, defaultRateLimit, maxBudget, conversationId, enableRateLimiting, enableCaching,
-                        enableCostTracking, task, isLazy, builtInSpecs, activeSpecs);
-                journalStore.markExecuted(conversationId, pauseEpoch, c.getCallId(), capUtf8(result, JOURNAL_RESULT_MAX_BYTES));
-                String envelope = amended != null ? amendedEnvelope(result) : result;
-                currentMessages.add(ToolExecutionResultMessage.from(rebuiltRequest(c), envelope));
-                clearedCallIds.add(c.getCallId());
-            } else {
-                // Duplicate claim — a prior attempt already ran (or crashed mid-tool).
-                var prior = journalStore.find(conversationId, pauseEpoch, c.getCallId());
-                if (prior.isPresent() && prior.get().status() == IHitlToolJournalStore.Status.EXECUTED) {
-                    // Replay the stored result — NEVER re-execute (no checkpoint re-fire).
-                    currentMessages.add(ToolExecutionResultMessage.from(rebuiltRequest(c), prior.get().resultCapped()));
-                    trace.add(Map.of("type", "hitl_replayed", "tool", c.getToolName(), "callId", c.getCallId()));
-                    clearedCallIds.add(c.getCallId());
-                } else {
-                    // EXECUTING (crash inside the tool) — honest at-most-once outcome.
-                    currentMessages.add(ToolExecutionResultMessage.from(rebuiltRequest(c),
-                            "{\"status\":\"EXECUTION_OUTCOME_UNKNOWN\",\"message\":\"a previous execution attempt was interrupted; "
-                                    + "it may or may not have taken effect — verify externally before retrying\"}"));
-                    auditOutcomeUnknown(memory, c);
-                    trace.add(Map.of("type", "hitl_outcome_unknown", "tool", c.getToolName(), "callId", c.getCallId()));
-                    clearedCallIds.add(c.getCallId());
-                }
-            }
-        }
-
-        // ── Step 4: continue the SAME loop from the next iteration (budget-continuous)
-        // ──
-        // The gate resolves the SAME way the live path did (LlmTask.executeTask): the
-        // cluster-wide eddi.hitl.tool.enabled kill-switch nulls the config (gate
-        // inert), otherwise task override else agent default, so NEW calls in the
-        // continuation re-gate → re-pause. Approved ids are pre-cleared so they are
-        // never re-gated if the model reissues them. Threading toolHitlEnabled here
-        // keeps the resume path from re-arming an approval flow an operator disabled.
-        ToolApprovalsConfig effectiveToolApprovals = null;
-        if (toolHitlEnabled) {
-            effectiveToolApprovals = task.getToolApprovals() != null
-                    ? task.getToolApprovals()
-                    : memory.getAgentToolApprovalsConfig();
-        }
-        int llmTaskIndex = batch.getLlmTaskIndex();
-
-        // Transcript cap: resumeToolLoop replays the ALREADY-serialized transcript
-        // via ChatTranscriptCodec#deserialize (no cap involved) and never threads the
-        // configured value in (LlmTask does not resolve it for the resume path). If
-        // this continuation re-gates a fresh call and re-pauses, the resulting batch
-        // is capped at the constant default here — a defensible, rare-path fallback
-        // rather than widening resumeToolLoop's signature for the primary knob, which
-        // governs the initial pause.
-        TokenUsage[] tokenHolder = new TokenUsage[1];
-        String response = runToolCallLoop(chatModel, currentMessages, activeSpecs, trace, batch.getIterationIndex() + 1,
-                setup, isLazy, task, memory, effectiveToolApprovals, llmTaskIndex, clearedCallIds, DEFAULT_TRANSCRIPT_MAX_BYTES, tokenHolder,
-                jsonPolicy);
-
-        // ── Step 5: merge the pre-pause trace with the resume trace ──
-        List<Map<String, Object>> mergedTrace = new ArrayList<>();
-        if (batch.getTraceSoFar() != null) {
-            mergedTrace.addAll(batch.getTraceSoFar());
-        }
-        mergedTrace.addAll(trace);
-
-        Map<String, Object> responseMetadata = new HashMap<>();
-        if (tokenHolder[0] != null) {
-            responseMetadata.put("tokenUsage", tokenUsageMap(tokenHolder[0]));
-        }
-        responseMetadata.put("toolCostUsd", toolCostDelta(conversationId, toolCostBefore));
-        return new ExecutionResult(response, mergedTrace, responseMetadata);
+        return toolLoopResumer.resumeToolLoop(chatModel, task, memory, batch, decision, toolHitlEnabled, jsonPolicy);
     }
 
     /** Journal-stored result cap (bytes) — matches the journal store's own cap. */
-    private static final int JOURNAL_RESULT_MAX_BYTES = 32_768;
+    static final int JOURNAL_RESULT_MAX_BYTES = 32_768;
 
-    /**
-     * Restores the active-spec surface on resume. For EAGER, every registered spec.
-     * For LAZY, exactly the specs that were active at pause time (by name), falling
-     * back to the LAZY-initial surface when the recorded names are absent.
-     */
+    /** @see ToolLoopResumer#restoreActiveSpecs */
     private static List<ToolSpecification> restoreActiveSpecs(ToolSetup setup, boolean isLazy, List<String> activatedToolNames) {
-        if (!isLazy) {
-            return setup.toolSpecs();
-        }
-        if (activatedToolNames == null || activatedToolNames.isEmpty()) {
-            return computeInitialActiveSpecs(setup, true);
-        }
-        Set<String> names = new HashSet<>(activatedToolNames);
-        List<ToolSpecification> restored = new ArrayList<>();
-        for (ToolSpecification spec : setup.toolSpecs()) {
-            if (names.contains(spec.name())) {
-                restored.add(spec);
-            }
-        }
-        return restored.isEmpty() ? computeInitialActiveSpecs(setup, true) : restored;
+        return ToolLoopResumer.restoreActiveSpecs(setup, isLazy, activatedToolNames);
     }
 
-    /**
-     * Rebuilds the base history exactly as a fresh turn would (the ONE sanctioned
-     * history rebuild on resume) and appends a reconstructed {@link AiMessage}
-     * carrying the batch's gated calls, so the appended results bind by call id.
-     * Intra-turn prior iterations are lost — accepted trade-off when the transcript
-     * cannot be restored. Uses the task's own (untemplated) system/prompt params;
-     * the orchestrator has no templating engine, and this degraded path only needs
-     * to carry the gated AiMessage so the appended results bind by call id.
-     */
+    /** @see ToolLoopResumer#fallbackRebuildMessages */
     private List<ChatMessage> fallbackRebuildMessages(LlmConfiguration.Task task, IConversationMemory memory,
                                                       PendingToolCallBatch batch) {
-        String systemMessage = "";
-        String prompt = null;
-        boolean includeFirstAgentMessage = true;
-        int logSizeLimit = task.getConversationHistoryLimit() != null ? task.getConversationHistoryLimit() : -1;
-        if (task.getParameters() != null) {
-            Object sys = task.getParameters().get("systemMessage");
-            if (sys != null) {
-                systemMessage = sys.toString();
-            }
-            Object p = task.getParameters().get("prompt");
-            if (p != null) {
-                prompt = p.toString();
-            }
-            Object inc = task.getParameters().get("includeFirstAgentMessage");
-            if (inc != null) {
-                includeFirstAgentMessage = Boolean.parseBoolean(inc.toString());
-            }
-        }
-
-        List<ChatMessage> base = conversationHistoryBuilder.buildMessages(memory, systemMessage, prompt, logSizeLimit,
-                includeFirstAgentMessage, null, 0);
-
-        List<ChatMessage> messages = new ArrayList<>(base);
-        List<ToolExecutionRequest> requests = new ArrayList<>();
-        for (PendingToolCallBatch.PendingToolCall c : batch.getCalls()) {
-            requests.add(rebuiltRequest(c));
-        }
-        messages.add(AiMessage.from(requests));
-        return messages;
+        return toolLoopResumer.fallbackRebuildMessages(task, memory, batch);
     }
 
     /**
@@ -746,80 +534,38 @@ class AgentOrchestrator implements IAgentOrchestrator {
     }
 
     /** Rebuilds a provider-safe request from a pending call (original raw args). */
+    /** @see ToolLoopResumer#rebuiltRequest */
     private static ToolExecutionRequest rebuiltRequest(PendingToolCallBatch.PendingToolCall c) {
-        return rebuiltRequest(c, c.getArgumentsRaw());
+        return ToolLoopResumer.rebuiltRequest(c);
     }
 
-    /**
-     * Rebuilds a request binding by the original call id with the given arguments.
-     */
+    /** @see ToolLoopResumer#rebuiltRequest */
     private static ToolExecutionRequest rebuiltRequest(PendingToolCallBatch.PendingToolCall c, String args) {
-        return ToolExecutionRequest.builder()
-                .id(c.getCallId())
-                .name(c.getToolName())
-                .arguments(args != null ? args : "")
-                .build();
+        return ToolLoopResumer.rebuiltRequest(c, args);
     }
 
-    /**
-     * Builds the reviewer-rejection envelope with the tool name and free-text note
-     * JSON-escaped via Jackson — NEVER string-concatenated with raw reviewer text.
-     */
+    /** @see ToolLoopResumer#rejectionEnvelope */
     private String rejectionEnvelope(String toolName, String note) {
-        Map<String, Object> envelope = new LinkedHashMap<>();
-        envelope.put("status", "REJECTED_BY_REVIEWER");
-        envelope.put("tool", toolName);
-        envelope.put("note", note != null ? note : "");
-        envelope.put("instruction", "The reviewer declined this action. Do not retry this exact call; "
-                + "explain the situation to the user and offer alternatives.");
-        return toJson(envelope);
+        return toolLoopResumer.rejectionEnvelope(toolName, note);
     }
 
-    /**
-     * Wraps an executed amended-args result so the model knows why it may differ.
-     */
+    /** @see ToolLoopResumer#amendedEnvelope */
     private String amendedEnvelope(String result) {
-        Map<String, Object> envelope = new LinkedHashMap<>();
-        envelope.put("status", "EXECUTED");
-        envelope.put("argsAmendedByReviewer", true);
-        envelope.put("result", result);
-        return toJson(envelope);
+        return toolLoopResumer.amendedEnvelope(result);
     }
 
     /**
      * Reusable Jackson mapper for approver/model-facing envelopes (escapes text).
      */
-    private static final ObjectMapper ENVELOPE_MAPPER = new ObjectMapper();
 
-    /**
-     * Jackson serialization for approver/model-facing JSON envelopes. All embedded
-     * reviewer/tool text is escaped by Jackson — NEVER string-concatenated.
-     */
+    /** @see ToolLoopResumer#toJson */
     private static String toJson(Object value) {
-        try {
-            return ENVELOPE_MAPPER.writeValueAsString(value);
-        } catch (com.fasterxml.jackson.core.JsonProcessingException e) {
-            // Fall back to a minimal, safe envelope rather than propagating — the
-            // resume must still complete. Effectively unreachable for the small maps
-            // serialized here.
-            LOGGER.warnf("HITL resume: JSON envelope serialization failed: %s", e.getMessage());
-            return "{\"status\":\"ERROR\",\"reason\":\"could not serialize result envelope\"}";
-        }
+        return ToolLoopResumer.toJson(value);
     }
 
-    /**
-     * Records an at-most-once outcome-unknown event. No lightweight
-     * {@code hitl.tool.*} audit collector is reachable from this task (the
-     * {@link ai.labs.eddi.engine.audit.model.AuditEntry} record is built by the
-     * LifecycleManager per-task with HMAC context we do not have here), so —
-     * exactly as the config-drift path does — this WARN-logs with a distinctive
-     * marker that operators can alert on. Package-private + overridable so tests
-     * can assert it fired.
-     */
+    /** @see ToolLoopResumer#auditOutcomeUnknown */
     void auditOutcomeUnknown(IConversationMemory memory, PendingToolCallBatch.PendingToolCall c) {
-        LOGGER.warnf("hitl.tool.outcome_unknown: approved tool '%s' (callId '%s') for conversation '%s' had an interrupted prior execution; "
-                + "outcome is unknown — verify externally before retrying.",
-                sanitize(c.getToolName()), sanitize(c.getCallId()), sanitize(memory.getConversationId()));
+        toolLoopResumer.auditOutcomeUnknown(memory, c);
     }
 
     /**
@@ -963,340 +709,40 @@ class AgentOrchestrator implements IAgentOrchestrator {
         }
     }
 
-    /**
-     * Computes the specs the LLM initially sees given a {@link ToolSetup}. For
-     * EAGER, that is every registered spec. For LAZY, only {@code discover_tools}
-     * plus the external (http/mcp/a2a) specs — the built-ins stay hidden until
-     * discovery activates them. Shared by the live loop and resume so both present
-     * an identical initial surface. Returns a fresh mutable list (LAZY activation
-     * mutates it in place).
-     */
+    /** @see ToolLoopRunner#computeInitialActiveSpecs */
     private static List<ToolSpecification> computeInitialActiveSpecs(ToolSetup setup, boolean isLazy) {
-        if (!isLazy) {
-            return setup.toolSpecs();
-        }
-        // External = every registered spec whose name is not a built-in spec name.
-        Set<String> builtInNames = new HashSet<>();
-        for (ToolSpecification spec : setup.builtInSpecs()) {
-            builtInNames.add(spec.name());
-        }
-        List<ToolSpecification> activeSpecs = new ArrayList<>();
-        int externalCount = 0;
-        for (ToolSpecification spec : setup.toolSpecs()) {
-            if ("discover_tools".equals(spec.name())) {
-                activeSpecs.add(spec);
-            } else if (!builtInNames.contains(spec.name())) {
-                activeSpecs.add(spec);
-                externalCount++;
-            }
-        }
-        LOGGER.infof("LAZY mode: presenting %d specs initially (discover_tools + %d external)",
-                activeSpecs.size(), externalCount);
-        return activeSpecs;
+        return ToolLoopRunner.computeInitialActiveSpecs(setup, isLazy);
     }
 
-    /**
-     * Executes the tool-calling loop using direct ChatModel API against a pre-built
-     * {@link ToolSetup} (shared with the resume path).
-     */
+    /** @see ToolLoopRunner#executeWithTools */
     private ExecutionResult executeWithTools(ChatModel chatModel, String systemMessage, List<ChatMessage> chatMessages, ToolSetup setup,
                                              LlmConfiguration.Task task, IConversationMemory memory,
                                              ToolApprovalsConfig effectiveToolApprovals, int llmTaskIndex, int transcriptMaxBytes,
                                              JsonResponseFormatPolicy jsonPolicy)
             throws LifecycleException {
-
-        Map<String, ToolExecutor> toolExecutors = setup.toolExecutors();
-        Map<String, String> toolSources = setup.toolSources();
-        List<ToolSpecification> builtInSpecs = setup.builtInSpecs();
-
-        boolean isLazy = task.getToolLoadingStrategy() == LlmConfiguration.ToolLoadingStrategy.LAZY;
-
-        // Active specs: what the LLM currently sees (LAZY starts narrow).
-        List<ToolSpecification> activeSpecs = computeInitialActiveSpecs(setup, isLazy);
-
-        // Build message list with system message if provided
-        List<ChatMessage> messages = new ArrayList<>();
-        if (!isNullOrEmpty(systemMessage)) {
-            messages.add(SystemMessage.from(systemMessage));
-        }
-        messages.addAll(chatMessages);
-
-        // Trace for tool calls
-        List<Map<String, Object>> trace = new ArrayList<>();
-
-        // Live path: iteration loop starts at 0, no pre-cleared call ids.
-        String conversationId = memory != null ? memory.getConversationId() : null;
-        double toolCostBefore = conversationToolCost(conversationId);
-        TokenUsage[] tokenHolder = new TokenUsage[1];
-        String response = runToolCallLoop(chatModel, messages, activeSpecs, trace, 0,
-                setup, isLazy, task, memory, effectiveToolApprovals, llmTaskIndex, Set.of(), transcriptMaxBytes, tokenHolder, jsonPolicy);
-
-        Map<String, Object> responseMetadata = new HashMap<>();
-        if (tokenHolder[0] != null) {
-            responseMetadata.put("tokenUsage", tokenUsageMap(tokenHolder[0]));
-        }
-        responseMetadata.put("toolCostUsd", toolCostDelta(conversationId, toolCostBefore));
-        return new ExecutionResult(response, trace, responseMetadata);
+        return toolLoopRunner.executeWithTools(chatModel, systemMessage, chatMessages, setup, task, memory,
+                effectiveToolApprovals, llmTaskIndex, transcriptMaxBytes, jsonPolicy);
     }
 
-    /**
-     * Accumulated tool cost for a conversation, or {@code 0.0} when nothing has
-     * been tracked for it yet — {@link ToolCostTracker#getConversationCosts}
-     * returns null for an untracked conversation.
-     */
+    /** @see ToolLoopRunner#conversationToolCost */
     private double conversationToolCost(String conversationId) {
-        if (conversationId == null || toolExecutionService == null) {
-            return 0.0;
-        }
-        ToolCostTracker tracker = toolExecutionService.getCostTracker();
-        if (tracker == null) {
-            return 0.0;
-        }
-        ToolCostTracker.ConversationCostMetrics metrics = tracker.getConversationCosts(conversationId);
-        return metrics != null ? metrics.getTotalCost() : 0.0;
+        return toolLoopRunner.conversationToolCost(conversationId);
     }
 
-    /**
-     * Dollar cost the tools of THIS model call added, as the difference between two
-     * snapshots of the conversation total. Clamped at zero because
-     * {@link ToolCostTracker#resetConversation} can fire between the snapshots and
-     * would otherwise yield a negative "cost".
-     */
+    /** @see ToolLoopRunner#toolCostDelta */
     private double toolCostDelta(String conversationId, double costBefore) {
-        return Math.max(0.0, conversationToolCost(conversationId) - costBefore);
+        return toolLoopRunner.toolCostDelta(conversationId, costBefore);
     }
 
-    /**
-     * The single shared tool-calling loop, wrapped in
-     * {@link AgentExecutionHelper#executeWithRetry}. Used by the live path (start
-     * iteration 0, empty {@code clearedCallIds}) and by {@link #resumeToolLoop}
-     * (start iteration {@code batch.getIterationIndex()+1}, {@code clearedCallIds}
-     * = the human-approved ids so they are never re-gated). A fresh gated batch
-     * throws {@link ToolApprovalRequiredException} to re-pause — the retry guard
-     * lets it escape unchanged.
-     *
-     * @param initialMessages
-     *            the message list the loop starts from (defensively copied inside);
-     *            for resume this already carries the replayed transcript + the
-     *            verdict-applied tool results
-     * @param startIteration
-     *            first loop index — carries budget continuity across a pause
-     * @param clearedCallIds
-     *            call ids a human already approved this pause; never re-gated
-     * @param transcriptMaxBytes
-     *            the configured cap (bytes) for freezing the transcript into a
-     *            {@link PendingToolCallBatch} if this iteration re-pauses
-     * @param jsonPolicy
-     *            decides, per request, whether {@code ResponseFormat.JSON} is set;
-     *            resolved against whether THAT request carries tool specifications
-     */
+    /** @see ToolLoopRunner#runToolCallLoop */
     private String runToolCallLoop(ChatModel chatModel, List<ChatMessage> initialMessages, List<ToolSpecification> activeSpecs,
                                    List<Map<String, Object>> trace, int startIteration, ToolSetup setup, boolean isLazy,
                                    LlmConfiguration.Task task, IConversationMemory memory, ToolApprovalsConfig effectiveToolApprovals,
                                    int llmTaskIndex, Set<String> clearedCallIds, int transcriptMaxBytes, TokenUsage[] tokenHolder,
                                    JsonResponseFormatPolicy jsonPolicy)
             throws LifecycleException {
-
-        Map<String, ToolExecutor> toolExecutors = setup.toolExecutors();
-        Map<String, String> toolSources = setup.toolSources();
-        Map<String, String> toolCanonicalNames = setup.toolCanonicalNames();
-        List<ToolSpecification> builtInSpecs = setup.builtInSpecs();
-
-        boolean enableRateLimiting = task.getEnableRateLimiting() != null ? task.getEnableRateLimiting() : true;
-        boolean enableCaching = task.getEnableToolCaching() != null ? task.getEnableToolCaching() : true;
-        boolean enableCostTracking = task.getEnableCostTracking() != null ? task.getEnableCostTracking() : true;
-        int defaultRateLimit = task.getDefaultRateLimit() != null ? task.getDefaultRateLimit() : 100;
-        Map<String, Integer> toolRateLimits = task.getToolRateLimits();
-        Double maxBudget = task.getMaxBudgetPerConversation();
-        String conversationId = memory.getConversationId();
-
-        // In-turn tool-context ceiling. Resolved once per loop: <= 0 disables the
-        // guard entirely (and skips estimator construction with it).
-        int toolContextBudget = task.getMaxToolContextTokens() != null
-                ? task.getMaxToolContextTokens()
-                : DEFAULT_MAX_TOOL_CONTEXT_TOKENS;
-        TokenCountEstimator toolContextEstimator = toolContextBudget > 0 ? toolContextBudgetGuard.resolveToolContextEstimator(task) : null;
-
-        return AgentExecutionHelper.executeWithRetry(() -> {
-            // A retry REPLAYS this whole lambda, so anything the previous attempt
-            // accumulated must be discarded — otherwise a turn that retried once
-            // reports (and bills) roughly double the tokens it actually used.
-            tokenHolder[0] = null;
-            List<ChatMessage> currentMessages = new ArrayList<>(initialMessages);
-            // Per-message token memo, local to this attempt: messages are immutable and
-            // re-counted on every iteration, so without it a 10-iteration loop tokenizes
-            // the first tool result ten times. Lives on the call stack — the orchestrator
-            // is a shared singleton and must stay stateless.
-            Map<ChatMessage, Integer> toolContextTokenMemo = new IdentityHashMap<>();
-            int maxIterations = task.getMaxToolIterations() != null ? task.getMaxToolIterations() : 10;
-
-            // Engine-enforced counterweight: strict mode caps iterations
-            var counterweight = task.getCounterweight();
-            if (counterweight != null && counterweight.isEnabled()
-                    && "strict".equalsIgnoreCase(counterweight.getLevel())) {
-                int strictCap = 5;
-                if (maxIterations > strictCap) {
-                    LOGGER.infof("Counterweight strict mode: capping tool iterations from %d to %d",
-                            maxIterations, strictCap);
-                    maxIterations = strictCap;
-                }
-            }
-
-            for (int i = startIteration; i < maxIterations; i++) {
-                // Cooperative cancellation: if this step was interrupted (e.g. a cascade
-                // per-step timeout called future.cancel(true)), stop before issuing another
-                // model call — avoids launching further side-effectful tools on a step whose
-                // result will be discarded. Thread.interrupted() also CLEARS the flag so it
-                // cannot leak to any later work on this thread.
-                if (Thread.interrupted()) {
-                    throw new LifecycleException("Agent execution cancelled (interrupted)");
-                }
-
-                // Keep the accumulated tool traffic inside its aggregate ceiling BEFORE
-                // the request is built — this is the only point at which the loop knows
-                // everything it is about to send. Without it the list only ever grows and
-                // a tool-heavy turn hard-fails mid-loop on a provider context-window 400.
-                if (toolContextEstimator != null) {
-                    enforceToolContextBudget(currentMessages, toolContextBudget, toolContextEstimator,
-                            toolContextTokenMemo, trace, conversationId);
-                }
-
-                ChatRequest.Builder requestBuilder = ChatRequest.builder().messages(currentMessages);
-
-                boolean toolsInRequest = !activeSpecs.isEmpty();
-                if (toolsInRequest) {
-                    requestBuilder.toolSpecifications(activeSpecs);
-                }
-
-                // API-level JSON, decided against THIS request's tool surface. Baking it
-                // into the model instead would put it on every request the cached model
-                // ever serves — the Gemini 400 documented in docs/changelog.md.
-                if (jsonPolicy != null) {
-                    var responseFormat = jsonPolicy.resolve(toolsInRequest);
-                    if (responseFormat != null) {
-                        requestBuilder.responseFormat(responseFormat);
-                    }
-                }
-
-                ChatRequest chatRequest = requestBuilder.build();
-
-                ChatResponse chatResponse = chatModel.chat(chatRequest);
-                AiMessage aiMessage = normalizeToolCallIds(chatResponse.aiMessage(), effectiveToolApprovals);
-                currentMessages.add(aiMessage);
-
-                // Accumulate token usage for cost/observability reporting.
-                if (chatResponse.metadata() != null && chatResponse.metadata().tokenUsage() != null) {
-                    tokenHolder[0] = sumTokens(tokenHolder[0], chatResponse.metadata().tokenUsage());
-                }
-
-                if (aiMessage.hasToolExecutionRequests()) {
-                    // === Tool-approval gate (tool-level HITL) ===
-                    // Split the batch into gated (require human approval) and allowed
-                    // calls. clearedCallIds carries the human-approved ids on resume so
-                    // they are never re-gated. Inert when effectiveToolApprovals is
-                    // null/empty — byte-identical to the pre-HITL path.
-                    var gateResult = toolApprovalGate.classify(aiMessage.toolExecutionRequests(), toolSources, setup.toolEndpoints(),
-                            effectiveToolApprovals, clearedCallIds);
-
-                    if (!gateResult.gated().isEmpty()) {
-                        int pausesSoFar = readToolPauseCount(memory);
-                        if (pausesSoFar >= maxPausesPerTurn(effectiveToolApprovals)) {
-                            // Fail-closed: the pause budget for this turn is spent —
-                            // do not pause and do not execute the gated calls; tell
-                            // the model to stop asking so the loop can still finish
-                            // with the ungated results.
-                            for (ToolExecutionRequest gatedReq : gateResult.gated()) {
-                                currentMessages.add(ToolExecutionResultMessage.from(gatedReq,
-                                        "{\"status\":\"DENIED\",\"reason\":\"approval-pause limit for this turn reached; do not retry\"}"));
-                                Map<String, Object> capStep = new HashMap<>();
-                                capStep.put("type", "tool_error");
-                                capStep.put("tool", gatedReq.name());
-                                capStep.put("error", "hitl_pause_cap");
-                                trace.add(capStep);
-                            }
-                            // Task 10 — the pause-cap guard fired: record it as a guard
-                            // (metric + audit) HERE where the audit context lives (the
-                            // memory's audit collector, set by ConversationService on the
-                            // say/resume paths). AgentOrchestrator is not CDI and has no
-                            // MeterRegistry, so the metric uses the Micrometer global
-                            // registry (same idiom as LifecycleManager). Emitted once per
-                            // capped batch, carrying the gated fingerprint.
-                            recordPauseCapGuard(memory, fingerprint(gateResult.gated()));
-                            // ungated calls still execute below
-                        } else {
-                            // 1) execute the ungated calls of this batch normally
-                            for (ToolExecutionRequest allowedReq : gateResult.allowed()) {
-                                executeSingleToolCall(allowedReq, memory, currentMessages, trace, toolExecutors,
-                                        toolRateLimits, toolCanonicalNames, defaultRateLimit, maxBudget, conversationId,
-                                        enableRateLimiting, enableCaching, enableCostTracking, task, isLazy, builtInSpecs, activeSpecs);
-                            }
-                            // Abandoned-thread guard: a cascade step that timed out (or
-                            // the agentTimeout watchdog on the live path) cancels the future
-                            // via cancel(true), interrupting THIS thread — but it keeps
-                            // running to here on the shared live memory while the caller has
-                            // already moved on. Committing pause state now would leave a
-                            // stale batch on a conversation the caller abandoned (self-heals
-                            // at next turn start) or overwrite a later step's real pause.
-                            // Abort WITHOUT mutating shared memory: throw the interrupted
-                            // signal instead of the pause. NOT ToolApprovalRequiredException,
-                            // which would commit a pause with a null batch. The guarantee
-                            // here is only that this abandoned thread never writes stale/
-                            // overwriting pause state; the resulting terminal state is
-                            // fail-safe either way — a cascade discards the Future's exception
-                            // entirely, and on the live path AgentExecutionHelper.executeWithRetry
-                            // re-wraps this into a plain LifecycleException so the turn settles
-                            // to ERROR (where the watchdog fired it already persisted
-                            // EXECUTION_INTERRUPTED). Both are recoverable by the next say.
-                            if (Thread.currentThread().isInterrupted()) {
-                                throw new LifecycleException.LifecycleInterruptedException(
-                                        "Tool-approval pause abandoned: executing thread was interrupted before commit");
-                            }
-                            // 2) resolve the per-tool friction rule (iteration 1). Done HERE,
-                            // once, because this is the last place the gated calls' endpoints
-                            // exist — the persisted batch keeps names and sources only, so an
-                            // endpoint-addressed rule could never be re-matched downstream.
-                            var ruleByCallId = ToolApprovalRules.matchByCallId(gateResult.gated(), toolSources,
-                                    setup.toolEndpoints(), effectiveToolApprovals);
-                            var governingRule = ToolApprovalRules.governing(ruleByCallId.values());
-                            recordRuleMatches(ruleByCallId.values());
-                            // 3) snapshot + persist the pending batch, then abort the loop
-                            PendingToolCallBatch batch = buildPendingBatch(currentMessages, gateResult, task, memory,
-                                    i, activatedToolNames(isLazy, activeSpecs), trace, pausesSoFar + 1, llmTaskIndex,
-                                    toolSources, effectiveToolApprovals, transcriptMaxBytes, ruleByCallId, governingRule);
-                            memory.setHitlPendingToolCalls(batch);
-                            incrementToolPauseCount(memory, pausesSoFar);
-                            throw new ToolApprovalRequiredException(
-                                    buildPauseReason(effectiveToolApprovals, gateResult, governingRule), batch);
-                        }
-                    }
-
-                    // Execute the allowed calls (was: aiMessage.toolExecutionRequests()).
-                    // When the gate is inert, gateResult.allowed() == the full batch.
-                    for (ToolExecutionRequest toolRequest : gateResult.allowed()) {
-                        // Cooperative cancellation before each (potentially side-effectful) tool.
-                        // Thread.interrupted() clears the flag so it cannot leak to later work.
-                        if (Thread.interrupted()) {
-                            throw new LifecycleException("Agent execution cancelled (interrupted) before tool: " + toolRequest.name());
-                        }
-
-                        executeSingleToolCall(toolRequest, memory, currentMessages, trace, toolExecutors,
-                                toolRateLimits, toolCanonicalNames, defaultRateLimit, maxBudget, conversationId,
-                                enableRateLimiting, enableCaching, enableCostTracking, task, isLazy, builtInSpecs, activeSpecs);
-                    }
-                } else {
-                    return aiMessage.text();
-                }
-            }
-
-            // Loop exhausted its iteration budget. The last message is usually the
-            // model's final AiMessage; on resume with a spent budget it may instead be
-            // a verdict-applied tool result — guard the cast either way.
-            ChatMessage last = currentMessages.get(currentMessages.size() - 1);
-            if (last instanceof AiMessage aiLast) {
-                return aiLast.text() != null ? aiLast.text() : "Max tool iterations reached";
-            }
-            return "Max tool iterations reached";
-        }, task, "Agent execution");
+        return toolLoopRunner.runToolCallLoop(chatModel, initialMessages, activeSpecs, trace, startIteration, setup, isLazy,
+                task, memory, effectiveToolApprovals, llmTaskIndex, clearedCallIds, transcriptMaxBytes, tokenHolder, jsonPolicy);
     }
 
     // ─── In-turn tool-context budget (D6b) — extracted to ToolContextBudget (R2
@@ -1320,16 +766,7 @@ class AgentOrchestrator implements IAgentOrchestrator {
         return ToolContextBudget.tokenUsageMap(usage);
     }
 
-    /**
-     * Executes a single tool call through the full per-request pipeline:
-     * auto-checkpoint, trace entry, per-conversation budget check, tenant cost
-     * budget check, executor dispatch (via {@link ToolExecutionService} for rate
-     * limiting/caching/cost tracking), response truncation, result trace, and LAZY
-     * activation. Extracted verbatim from the live loop so the live path and Task
-     * 9's resume path share ONE copy.
-     * <p>
-     * Package-private for direct unit testing of the gate.
-     */
+    /** @see ToolLoopRunner#executeSingleToolCall */
     void executeSingleToolCall(ToolExecutionRequest toolRequest, IConversationMemory memory,
                                List<ChatMessage> currentMessages, List<Map<String, Object>> trace,
                                Map<String, ToolExecutor> toolExecutors, Map<String, Integer> toolRateLimits,
@@ -1338,24 +775,12 @@ class AgentOrchestrator implements IAgentOrchestrator {
                                boolean enableRateLimiting, boolean enableCaching, boolean enableCostTracking,
                                LlmConfiguration.Task task, boolean isLazy,
                                List<ToolSpecification> builtInSpecs, List<ToolSpecification> activeSpecs) {
-        // Live path: run the full pipeline, then append the raw result verbatim.
-        String toolResult = executeSingleToolCallResult(toolRequest, memory, trace, toolExecutors, toolRateLimits,
+        toolLoopRunner.executeSingleToolCall(toolRequest, memory, currentMessages, trace, toolExecutors, toolRateLimits,
                 toolCanonicalNames, defaultRateLimit, maxBudget, conversationId, enableRateLimiting, enableCaching,
                 enableCostTracking, task, isLazy, builtInSpecs, activeSpecs);
-        currentMessages.add(ToolExecutionResultMessage.from(toolRequest, toolResult));
     }
 
-    /**
-     * Runs the full per-request pipeline (auto-checkpoint, trace entry,
-     * per-conversation budget check, tenant cost budget check, executor dispatch
-     * via {@link ToolExecutionService}, response truncation, result trace, LAZY
-     * activation) and RETURNS the resulting tool-result string WITHOUT appending it
-     * to any message list. The live loop appends the raw result; the resume path
-     * needs the string to journal it and to build the amended-args envelope, so it
-     * appends its own (possibly-enveloped) message. This is the single shared
-     * per-request pipeline for both paths — the auto-checkpoint fires here (and
-     * only here) so replayed/outcome-unknown resume calls never re-checkpoint.
-     */
+    /** @see ToolLoopRunner#executeSingleToolCallResult */
     String executeSingleToolCallResult(ToolExecutionRequest toolRequest, IConversationMemory memory,
                                        List<Map<String, Object>> trace,
                                        Map<String, ToolExecutor> toolExecutors, Map<String, Integer> toolRateLimits,
@@ -1364,153 +789,20 @@ class AgentOrchestrator implements IAgentOrchestrator {
                                        boolean enableRateLimiting, boolean enableCaching, boolean enableCostTracking,
                                        LlmConfiguration.Task task, boolean isLazy,
                                        List<ToolSpecification> builtInSpecs, List<ToolSpecification> activeSpecs) {
-        // Auto-checkpoint before tool execution (Wave 4)
-        if (memorySnapshotService != null) {
-            try {
-                memorySnapshotService.createCheckpoint(
-                        memory, "before_tool:" + toolRequest.name(),
-                        AgentOrchestrator.class.getSimpleName());
-            } catch (Exception cpEx) {
-                LOGGER.warnf("Auto-checkpoint failed before tool '%s': %s",
-                        toolRequest.name(), cpEx.getMessage());
-            }
-        }
-
-        Map<String, Object> callStep = new HashMap<>();
-        callStep.put("type", "tool_call");
-        callStep.put("tool", toolRequest.name());
-        callStep.put("arguments", toolRequest.arguments());
-        trace.add(callStep);
-
-        // Check per-conversation TOOL budget before executing tool.
-        //
-        // Enforcement is opt-in: enforceBudget on the task wins, otherwise
-        // BUDGET_ENFORCE_DEFAULT (false unless eddi.tools.budget.enforce-by-default
-        // says so). Built-in tools priced at $0.00 until this release, so enforcing
-        // by default would make those ceilings bind for the first time on upgrade.
-        // A ceiling left unenforced is named once in a startup-style WARN rather
-        // than passing silently. Cost tracking itself is unaffected by the flag.
-        warnAboutUnenforcedBudgets(task);
-        boolean enforceBudget = task.getEnforceBudget() != null ? task.getEnforceBudget() : BUDGET_ENFORCE_DEFAULT;
-        if (enforceBudget && maxBudget != null && conversationId != null
-                && !toolExecutionService.getCostTracker().isWithinBudget(conversationId, maxBudget)) {
-            String budgetError = "Budget exceeded for conversation " + conversationId;
-            LOGGER.warn(sanitize(budgetError));
-
-            Map<String, Object> budgetStep = new HashMap<>();
-            budgetStep.put("type", "tool_error");
-            budgetStep.put("tool", toolRequest.name());
-            budgetStep.put("error", budgetError);
-            trace.add(budgetStep);
-
-            return "Error: " + budgetError;
-        }
-
-        // Check tenant-level monthly cost budget (MCP governance)
-        if (tenantQuotaService != null) {
-            var costCheck = tenantQuotaService.checkCostBudget(tenantQuotaService.getDefaultTenantId());
-            if (!costCheck.allowed()) {
-                LOGGER.warnf("Tenant cost budget exceeded during tool call: %s", costCheck.reason());
-
-                Map<String, Object> quotaStep = new HashMap<>();
-                quotaStep.put("type", "tool_error");
-                quotaStep.put("tool", toolRequest.name());
-                quotaStep.put("error", costCheck.reason());
-                trace.add(quotaStep);
-
-                return "Error: " + costCheck.reason();
-            }
-        }
-
-        // Execute through ToolExecutionService for rate limiting, caching, cost
-        // tracking
-        ToolExecutor executor = toolExecutors.get(toolRequest.name());
-        String toolResult;
-        if (executor != null) {
-            // A built-in is DISPATCHED under its @Tool method name ("calculate") but
-            // CONFIGURED under its whitelist slug ("calculator"). Resolve the slug once,
-            // here, and hand both names down: only the price and the cache TTL are looked
-            // up by slug, everything that identifies the individual call is not.
-            String canonicalName = ToolNameResolver.canonical(toolRequest.name(), toolCanonicalNames);
-            int rateLimit = resolveRateLimit(toolRateLimits, toolRequest.name(), canonicalName, defaultRateLimit);
-            Double priceOverride = resolveOverride(task.getToolPricing(), toolRequest.name(), canonicalName);
-
-            // Partition the tool-result cache by identity, so one user's result can never
-            // be served back to another. A null tag means no usable identity was
-            // available; ToolExecutionService then bypasses the cache entirely. Both
-            // names go in: toolCacheScopes shares its key vocabulary with toolRateLimits
-            // and toolPricing, so a slug-keyed narrowing override has to bind here too.
-            String cacheScopeTag = ToolCacheService.resolveScopeTag(toolRequest.name(), canonicalName, task.getToolCacheScopes(),
-                    task.getDefaultToolCacheScope(), memory != null ? memory.getUserId() : null, conversationId);
-
-            var invocation = new ToolInvocation(toolRequest.name(), canonicalName, priceOverride);
-            toolResult = toolExecutionService.executeToolWrapped(invocation, toolRequest.arguments(), cacheScopeTag, conversationId,
-                    () -> executor.execute(toolRequest, null), enableRateLimiting, enableCaching, enableCostTracking, rateLimit);
-        } else {
-            toolResult = "Error: Tool '" + toolRequest.name() + "' not found";
-        }
-
-        // Apply response truncation (MCP governance)
-        toolResult = toolResponseTruncator.truncateIfNeeded(
-                toolRequest.name(), toolResult, task.getToolResponseLimits(),
-                task.getType(), task.getParameters());
-
-        Map<String, Object> resultStep = new HashMap<>();
-        resultStep.put("type", "tool_result");
-        resultStep.put("tool", toolRequest.name());
-        resultStep.put("result", toolResult);
-        trace.add(resultStep);
-
-        // LAZY mode: after discover_tools returns, activate the matching built-in specs
-        if (isLazy && "discover_tools".equals(toolRequest.name())) {
-            activateDiscoveredTools(toolResult, builtInSpecs, activeSpecs);
-        }
-
-        return toolResult;
+        return toolLoopRunner.executeSingleToolCallResult(toolRequest, memory, trace, toolExecutors, toolRateLimits,
+                toolCanonicalNames, defaultRateLimit, maxBudget, conversationId, enableRateLimiting, enableCaching,
+                enableCostTracking, task, isLazy, builtInSpecs, activeSpecs);
     }
 
-    /**
-     * Resolves the per-minute rate limit for one call: an entry keyed on the
-     * dispatch name wins, then the canonical slug, then the task default.
-     *
-     * <p>
-     * Dispatch-name-first keeps a config that pins a single method
-     * ({@code {"searchNews": 5}}) more specific than one that covers the whole tool
-     * ({@code {"websearch": 30}}), and is what makes the documented slug form bind
-     * at all — a built-in's dispatch name never equals its slug.
-     * </p>
-     *
-     * <p>
-     * Note that only the LIMIT is slug-resolved; the bucket itself lives in
-     * {@code ToolRateLimiter} under the dispatch name. {@code {"websearch": 30}}
-     * therefore grants {@code searchWeb}, {@code searchNews} and
-     * {@code searchWikipedia} 30 calls/min <em>each</em>, not 30 between them.
-     * </p>
-     */
+    /** @see ToolLoopRunner#resolveRateLimit */
     static int resolveRateLimit(Map<String, Integer> toolRateLimits, String dispatchName, String canonicalName,
                                 int defaultRateLimit) {
-        if (toolRateLimits == null) {
-            return defaultRateLimit;
-        }
-        Integer limit = toolRateLimits.get(dispatchName);
-        if (limit == null) {
-            limit = toolRateLimits.get(canonicalName);
-        }
-        return limit != null ? limit : defaultRateLimit;
+        return ToolLoopRunner.resolveRateLimit(toolRateLimits, dispatchName, canonicalName, defaultRateLimit);
     }
 
-    /**
-     * Resolves an operator price override for one call, dispatch name before
-     * canonical slug (same precedence as {@link #resolveRateLimit}). Returns
-     * {@code null} when no override applies, leaving the default price table in
-     * charge.
-     */
+    /** @see ToolLoopRunner#resolveOverride */
     static Double resolveOverride(Map<String, Double> toolPricing, String dispatchName, String canonicalName) {
-        if (toolPricing == null) {
-            return null;
-        }
-        Double price = toolPricing.get(dispatchName);
-        return price != null ? price : toolPricing.get(canonicalName);
+        return ToolLoopRunner.resolveOverride(toolPricing, dispatchName, canonicalName);
     }
 
     // ─── Tool-approval gate helpers ───
@@ -1620,48 +912,11 @@ class AgentOrchestrator implements IAgentOrchestrator {
         gateSupport.recordPauseCapGuard(memory, fingerprint);
     }
 
-    /**
-     * Parses the discover_tools JSON result and activates matching built-in tool
-     * specs so the LLM can call them on subsequent iterations.
-     */
+    /** @see ToolLoopRunner#activateDiscoveredTools */
     private void activateDiscoveredTools(String discoverResult,
                                          List<ToolSpecification> builtInSpecs,
                                          List<ToolSpecification> activeSpecs) {
-        try {
-            ObjectMapper mapper = new ObjectMapper();
-            JsonNode root = mapper.readTree(discoverResult);
-            JsonNode toolsNode = root.get("tools");
-            if (toolsNode == null || !toolsNode.isArray()) {
-                return;
-            }
-
-            Set<String> discoveredNames = new HashSet<>();
-            for (JsonNode tool : toolsNode) {
-                if (tool.has("name")) {
-                    discoveredNames.add(tool.get("name").asText());
-                }
-            }
-
-            // Add matching specs (skip discover_tools itself and already-active specs)
-            Set<String> activeNames = new HashSet<>();
-            for (ToolSpecification spec : activeSpecs) {
-                activeNames.add(spec.name());
-            }
-
-            int activated = 0;
-            for (ToolSpecification spec : builtInSpecs) {
-                if (discoveredNames.contains(spec.name()) && !activeNames.contains(spec.name())) {
-                    activeSpecs.add(spec);
-                    activated++;
-                }
-            }
-
-            LOGGER.infof("LAZY activation: %d tools activated from discovery (%s)",
-                    activated, discoveredNames);
-        } catch (Exception e) {
-            LOGGER.warnf("Failed to parse discover_tools result for LAZY activation: %s",
-                    e.getMessage());
-        }
+        toolLoopRunner.activateDiscoveredTools(discoverResult, builtInSpecs, activeSpecs);
     }
 
     /**
